add SuggestionData
